// models/request/institutionInfoRequest.ts
export interface InstitutionInfoRequest {
    sendingInstitutionInfo_id: number;
    hei_id: string;
    universityDepartment_id: number;
    academicPersonnelName: string;
    academicPersonnelSurname: string;
    academicPersonnelEmail: string;
    administrativePersonnelName: string;
    administrativePersonnelSurname: string;
    administrativePersonnelEmail: string;
    phoneNumberE164: string;
    phoneNumberExt: string;
}